Moved Views from package dk.ule.oapenwb.base to dk.ule.oapenwb.entity for easier outsourcing into the new oapenwb-persistency project
