+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/Dict.java
@@ -11,7 +11,7 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import dk.ule.oapenwb.base.AppConfig;
 import dk.ule.oapenwb.base.RunMode;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.base.error.CodeException;
 import dk.ule.oapenwb.data.DataInitializer;
 import dk.ule.oapenwb.data.importer.FileImporter;
@@ -27,7 +27,6 @@ import dk.ule.oapenwb.rpc.DictSpring;
 import dk.ule.oapenwb.util.CurrentUser;
 import dk.ule.oapenwb.util.EmailUtil;
 import dk.ule.oapenwb.util.HibernateUtil;
-import io.grpc.StatusRuntimeException;
 import io.javalin.Javalin;
 import io.javalin.http.Context;
 import io.javalin.http.Handler;
