+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/Views.java
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: © 2022 Michael Köther <mkoether38@gmail.com>
 // SPDX-License-Identifier: AGPL-3.0-only
-package dk.ule.oapenwb.base;
+package dk.ule.oapenwb.entity;
 
 /**
  * The different views to be used in the @JsonView annotation are defined as subclasses of this Views class.
