+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/basedata/tlConfig/TypeLanguageConfig.java
@@ -4,7 +4,7 @@ package dk.ule.oapenwb.entity.content.basedata.tlConfig;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.entity.content.basedata.LexemeFormType;
 import dk.ule.oapenwb.logic.admin.generic.IEntity;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
