+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/lexemes/LexemeForm.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.entity.content.lexemes;
 
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
