+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/lexemes/MetaInfo.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.entity.content.lexemes;
 
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.entity.content.lexemes.lexeme.Variant;
 import lombok.Data;
 import lombok.NoArgsConstructor;
