+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/ui/UiTranslationSet.java
@@ -4,7 +4,7 @@ package dk.ule.oapenwb.entity.ui;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.logic.admin.generic.IEntity;
 import lombok.Data;
 import lombok.NoArgsConstructor;
