+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/syngroup/SGSearchResult.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.logic.admin.syngroup;
 
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.logic.admin.lexeme.sememe.LexemeSlimPlus;
 import lombok.Data;
