+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/config/BaseConfig.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.logic.config;
 
 import com.fasterxml.jackson.annotation.JsonView;
-import dk.ule.oapenwb.base.Views;
+import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.entity.content.basedata.*;
 import dk.ule.oapenwb.entity.ui.UiLanguage;
 import lombok.Data;
